refactor(Outcome) : 상태 객체 이동
