+++ src/main/kotlin/blackjack/domain/state/State.kt
@@ -2,7 +2,6 @@ package blackjack.domain.state
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
-import blackjack.domain.result.Outcome
 
 interface State {
     val cards: Cards
