+++ src/main/kotlin/blackjack/domain/state/inTurn/FirstTurn.kt
@@ -2,7 +2,7 @@ package blackjack.domain.state.inTurn
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
-import blackjack.domain.result.Outcome
+import blackjack.domain.state.Outcome
 import blackjack.domain.state.State
 import blackjack.domain.state.endTurn.BlackJack
