+++ src/test/kotlin/blackjack/domain/state/OutcomeTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.result
+package blackjack.domain.state
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardMark
